Clean up failing web-sockets
Use a channel based data structure for keeping track of the web sockets
that have been opened. Exposes an Add and a Remove channel that accepts
websocket pointers and will add or remove web sockets to the internal
array. Using channels, means that the data structure is thread safe, and
by removing the web sockets once we've received an error on them, means
that we won't continue to fire traffic at broken sockets.